easy_telegram_chat_id/requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
python-telegram-bot==21.7
sniffio==1.3.1
typing_extensions==4.12.2
